added euclidean algorithms and extended

+++ lib/primes/primes.h
@@ -1,8 +1,10 @@
 #ifndef PRIMES_H_INCLUDED
 #define PRIMES_H_INCLUDED
-#include <stdbool.h>
 
-  bool cryptics_primes_primality_test(unsigned long long);
-  bool cryptics_primes_coprimality_test(unsigned long long number_one, unsigned long long number_two);
+  #include <stdbool.h>
+  #include <stdint.h>
+
+  bool cryptics_primes_primality_test(uint64_t);
+  bool cryptics_primes_coprimality_test(uint64_t number_one, uint64_t number_two);
 
 #endif // PRIMES_H_INCLUDED
